feat!: release v4.0.0 — Bun toolchain, CLI, transcripts, full serverless coverage
This commit is the conventional-commit marker that drives semantic-release
to tag and publish v4.0.0. The diff itself clears two non-fatal warnings
from the rollup build; the v4 implementation work landed across the prior
commits on this branch.

### Build cleanup in this commit

1. Circular dependency: src/index.ts -> src/batch-verifier.ts -> src/index.ts
   index.ts re-exported verifyEmailBatch from batch-verifier.ts, which
   imported verifyEmail back from ./index. Benign at runtime but rollup
   correctly flagged the fragility. Extracted verifyEmail (+ private
   helpers) into src/verify-email.ts and the two list checks
   (isDisposableEmail, isFreeEmail) into their own files. src/index.ts is
   now a pure barrel; batch-verifier.ts imports verifyEmail directly from
   ./verify-email. Public API surface unchanged.

2. esbuild "import.meta is not available in target (es2019)" warning
   src/cli/index.ts uses import.meta.main for Bun's direct-execution
   check. The CLI bundle now uses target: 'es2020' (where import.meta is
   part of the spec); the main package bundles stay on es2019. Output is
   still CJS — Node treats import.meta as undefined there, which is what
   isDirectInvocation expects (require.main === module carries the CJS
   branch).

Build now logs zero warnings; all 12 bundle outputs produced cleanly.

### v4.0.0 release highlights

- New `email-validate` CLI binary (full pipeline, transcripts, JSON log
  files, exit codes, programmatic surface)
- Structured transcript capture across SMTP / MX / WHOIS / disposable /
  free / typo / name-detection
- Six serverless platform adapters: AWS Lambda, GCP Cloud Functions
  (2nd gen), Vercel (Edge + Node), Cloudflare Workers + Durable Objects,
  Netlify Functions, Azure Functions (v4)
- Built-in zero-dep DoHResolver — DNS-over-HTTPS for edge MX checks
  (Cloudflare 1.1.1.1 default; configurable; cf-doh-compatible)
- Migration from yarn + jest + ts-jest + eslint to Bun + bun:test +
  Biome (CI + release workflows updated)
- ~770 default tests + ~200 extras tests + 112 integration tests
- Zero `!` non-null assertions in src/ (now an AGENTS.md rule)

See CHANGELOG.md v4.0.0 entry for the full breakdown.

BREAKING CHANGE: Stricter email format validator now rejects local-part
characters that are RFC-5322-valid but virtually never seen in real
mailboxes. The local-part allow-list is now positive: [a-zA-Z0-9._+'-].
Addresses like `hey=mo@gmail.com`, `foo?bar@gmail.com`, `foo^bar@…`,
`foo!bar@…`, `foo$bar@…`, `foo&bar@…`, `foo|bar@…` are no longer
considered valid. Standard sub-addressing with +, dots, underscores,
hyphens, and apostrophes (o'brien@…) is unchanged.

BREAKING CHANGE: check-if-email-exists.ts and its companion fixtures are
no longer in the published src/ tree. They moved to
extras/check-if-email-exists/ and are now opt-in via
`bun run test:extras`. Imports change from
`@emailcheck/email-validator-js/check-if-email-exists` to
`@emailcheck/email-validator-js/extras/check-if-email-exists`.

BREAKING CHANGE: validateBatchEmailsField (used by serverless adapters)
now returns a discriminated union { ok: true; emails: string[] } |
{ ok: false; status; message } instead of a nullable error object. This
removed all non-null assertions from adapter code. Anyone re-implementing
this helper layer should mirror the new shape.

BREAKING CHANGE: Test layout moved from a flat __tests__/ to
__tests__/{unit,isolated,integration}/ subfolders. The test scripts
(test, test:unit, test:isolated, etc.) are unchanged externally, but
external tooling that targeted specific paths needs to update.

BREAKING CHANGE: Examples reorganized into topic folders. Files like
examples/smtp-usage.ts are now examples/smtp/usage.ts; full mapping in
examples/README.md.
